+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+##### 30-June-2018 **1.0.4**:
+* added new method throw
+* fixed container storage bug
+* fixed meta builder bug
+
 ##### 24-June-2018 **1.0.3**:
 * Added documentation in English